Add standalone boot parameters and BL2 binary files
bl2_bp.bin cannot be used in a SD card image. The bootrom expect the boot
parameters at sector 1, and the BL2 program at sector 8, so there is a gap of 6
sectors between the two. bl2_bp.bin has the boot parameters immediately followed
by BL2, so there is no such gap.

Add a separate bp.bin and bl2 files, so that they can be used in a WKS file to
generate a proper SD card image.

Signed-off-by: Hugo Villeneuve <hvilleneuve@dimonoff.com>